Add sliding animation, fade in descriptions
Figured out how to smoothly slide in the photos (in both directions!). Need to put animated
element inside a container that has "overflow:hidden" to have other page elements clip the contents of the animated image.

Also, to stop the effect of "image is drawn as a huge original size, then shrinks down, then animates", which looks very ugly and jarring, I needed to add "width=100%" to the image (and not give the original width and height to the <img> tag) to make the image render to 100% of the width of the parent container, which displays the correct behavior without the jarring resizing.

To possibly activate GPU rendering on mobile devices and make the
sliding animation much smoother, I need to activate the CSS properties
"will-change: transform" and use "translateX()" instead of using the top, right, bottom, or left margins for animation.