Refactor pnButtons update correctly in setState()
I didn't realize that setting pnButtons with an object will completely
replace the old pnButtons object in this.state, so that's why I was
having problems losing the 'this' binding for the mediaButtonClick
method. I was assuming that using the pnButtons object and only setting
particular properties will only update those specific properties, but
because I was grouping the properties into one pnButtons object, the
whole object was being replaced, so I was losing any properties that
existed before.

The way to fix this is to do a shallow merge with Object.assign() to
return a new object that keeps older properties while updating new
propreties that changed.

Thank you, Wallace!
https://github.com/lukwallace