Fix numberbar click to not make image disappear
The previous erroneous behavior was that if the user clicked on the same
index that was currently shown, the image would fade to almost 0%
opacity.

I had to develop new functionality where the App would detect
if the click came from the navbar. If true, then do not set the
imgLoaded flag to false.

The imgLoaded flag's normal behavior is to start a CSS animation to 0%
opacity for unloaded images until the network call is fully finished and
the image has been loaded to the DOM, where the imgLoaded flag would be
set to true, and this triggers a new class to be added that will start
a CSS animation to fade the image from 0% opacity to 100% opacity.